clonedriveuploader/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
google-auth-oauthlib
google-api-python-client
requests
python-dotenv
google-cloud-aiplatform
google-cloud-texttospeech
Pillow
moviepy
numpy
openai
assemblyai
google-auth-httplib2
google-api-core
google-cloud-storage
google-cloud-vision
google-cloud-videointelligence
google-cloud-translate
google-cloud-language
google-cloud-speech
diffusers
pydub
